Moved more forms to the form module.
